@Sean-Der 2019-03-05
This looks really good, thank you so much for digging into this!

Can you add tests for
* Closing with subs
* Writing with no subs

Then this is a merge from me, we can tag and start getting this out. Nice work :)